+++ packages/jest-cli/src/TestScheduler.js
@@ -14,6 +14,7 @@ import type {Reporter, Test} from 'types/TestRunner';
 
 import chalk from 'chalk';
 import {formatExecError} from 'jest-message-util';
+import {interopRequireDefault} from 'jest-util';
 import {
   addResult,
   buildFailureTestResult,
@@ -24,7 +25,7 @@ import DefaultReporter from './reporters/default_reporter';
 import exit from 'exit';
 import NotifyReporter from './reporters/notify_reporter';
 import ReporterDispatcher from './ReporterDispatcher';
-import snapshot from 'jest-snapshot';
+import {cleanup as snapshotCleanup, buildSnapshotResolver} from 'jest-snapshot';
 import SummaryReporter from './reporters/summary_reporter';
 import TestRunner from 'jest-runner';
 import TestWatcher from './TestWatcher';
@@ -144,10 +145,10 @@ export default class TestScheduler {
 
     const updateSnapshotState = () => {
       contexts.forEach(context => {
-        const status = snapshot.cleanup(
+        const status = snapshotCleanup(
           context.hasteFS,
           this._globalConfig.updateSnapshot,
-          snapshot.buildSnapshotResolver(context.config),
+          buildSnapshotResolver(context.config),
         );
 
         aggregatedResults.snapshot.filesRemoved += status.filesRemoved;
@@ -170,10 +171,12 @@ export default class TestScheduler {
     const testRunners = Object.create(null);
     contexts.forEach(({config}) => {
       if (!testRunners[config.runner]) {
+        const Runner: TestRunner = interopRequireDefault(
+          // $FlowFixMe: dynamic import
+          require(config.runner),
+        ).default;
         // $FlowFixMe
-        testRunners[config.runner] = new (require(config.runner): TestRunner)(
-          this._globalConfig,
-        );
+        testRunners[config.runner] = new Runner(this._globalConfig);
       }
     });