+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -6,16 +6,15 @@
  *
  */
 
+import crypto from 'crypto';
+import path from 'path';
 import {escapeStrForRegex} from 'jest-regex-util';
 import normalize from '../normalize';
 
-jest.mock('jest-resolve');
+import resolve from 'jest-resolve';
+import {DEFAULT_JS_PATTERN} from '../constants';
 
-jest.mock('path', () => jest.requireActual('path').posix);
-
-const crypto = require('crypto');
-const path = require('path');
-const DEFAULT_JS_PATTERN = require('../constants').DEFAULT_JS_PATTERN;
+jest.mock('jest-resolve').mock('path', () => jest.requireActual('path').posix);
 const DEFAULT_CSS_PATTERN = '^.+\\.(css)$';
 
 let root;
@@ -45,7 +44,7 @@ beforeEach(() => {
   expectedPathAbs = path.join(root, 'an', 'abs', 'path');
   expectedPathAbsAnother = path.join(root, 'another', 'abs', 'path');
 
-  require('jest-resolve').findNodeModule = findNodeModule;
+  resolve.findNodeModule = findNodeModule;
 });
 
 it('picks a name based on the rootDir', () => {
@@ -281,7 +280,7 @@ describe('roots', () => {
 describe('transform', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => name);
   });
 
@@ -309,7 +308,7 @@ describe('transform', () => {
 describe('haste', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => name);
   });
 
@@ -333,7 +332,7 @@ describe('haste', () => {
 describe('setupFilesAfterEnv', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name =>
       name.startsWith('/') ? name : '/root/path/foo' + path.sep + name,
     );
@@ -383,7 +382,7 @@ describe('setupTestFrameworkScriptFile', () => {
   beforeEach(() => {
     console.warn = jest.fn();
     consoleWarn = console.warn;
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name =>
       name.startsWith('/') ? name : '/root/path/foo' + path.sep + name,
     );
@@ -636,7 +635,7 @@ describe('testRunner', () => {
   });
 
   it('is overwritten by argv', () => {
-    const Resolver = require('jest-resolve');
+    const Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => name);
     const {options} = normalize(
       {
@@ -667,7 +666,7 @@ describe('coverageDirectory', () => {
 describe('testEnvironment', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => {
       if (['jsdom', 'jest-environment-jsdom'].includes(name)) {
         return `node_modules/${name}`;
@@ -721,7 +720,7 @@ describe('testEnvironment', () => {
 describe('babel-jest', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name =>
       name.indexOf('babel-jest') === -1
         ? path.sep + 'node_modules' + path.sep + name
@@ -805,7 +804,7 @@ describe('Upgrade help', () => {
     consoleWarn = console.warn;
     console.warn = jest.fn();
 
-    const Resolver = require('jest-resolve');
+    const Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => {
       if (name == 'bar/baz') {
         return '/node_modules/bar/baz';
@@ -954,7 +953,7 @@ describe('moduleDirectories', () => {
 
 describe('preset', () => {
   beforeEach(() => {
-    const Resolver = require('jest-resolve');
+    const Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => {
       if (name === 'react-native/jest-preset') {
         return '/node_modules/react-native/jest-preset.json';
@@ -1062,7 +1061,7 @@ describe('preset', () => {
   });
 
   test('searches for .json and .js preset files', () => {
-    const Resolver = require('jest-resolve');
+    const Resolver = require('jest-resolve').default;
 
     normalize(
       {
@@ -1157,7 +1156,7 @@ describe('preset', () => {
 describe('preset without setupFiles', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(
       name => path.sep + 'node_modules' + path.sep + name,
     );
@@ -1202,7 +1201,7 @@ describe('preset without setupFiles', () => {
 describe('runner', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => {
       if (['eslint', 'jest-runner-eslint', 'my-runner-foo'].includes(name)) {
         return `node_modules/${name}`;
@@ -1260,7 +1259,7 @@ describe('runner', () => {
 describe('watchPlugins', () => {
   let Resolver;
   beforeEach(() => {
-    Resolver = require('jest-resolve');
+    Resolver = require('jest-resolve').default;
     Resolver.findNodeModule = jest.fn(name => {
       if (
         ['typeahead', 'jest-watch-typeahead', 'my-watch-plugin'].includes(name)
@@ -1394,7 +1393,7 @@ describe('testPathPattern', () => {
       describe('win32', () => {
         beforeEach(() => {
           jest.mock('path', () => jest.requireActual('path').win32);
-          require('jest-resolve').findNodeModule = findNodeModule;
+          require('jest-resolve').default.findNodeModule = findNodeModule;
         });
 
         afterEach(() => {