+++ packages/jest-repl/src/cli/repl.js
@@ -17,6 +17,7 @@ declare var jest: Object;
 import path from 'path';
 import repl from 'repl';
 import vm from 'vm';
+import {interopRequireDefault} from 'jest-util';
 
 let transformer;
 
@@ -59,8 +60,8 @@ if (jestProjectConfig.transform) {
     }
   }
   if (transformerPath) {
-    /* $FlowFixMe */
-    transformer = require(transformerPath);
+    // $FlowFixMe
+    transformer = interopRequireDefault(require(transformerPath)).default;
     if (typeof transformer.process !== 'function') {
       throw new TypeError(
         'Jest: a transformer must export a `process` function.',