+++ types/SnapshotResolver.js
@@ -3,6 +3,7 @@
 import type {Path} from './Config';
 
 export type SnapshotResolver = {|
+  testPathForConsistencyCheck: string,
   resolveSnapshotPath(testPath: Path): Path,
   resoveTestPath(snapshotPath: Path): Path,
 |};